mvvm 0.9

+++ SpawmetDatabaseWPF/DeleteMachineWindow.xaml.cs
@@ -1,6 +1,7 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.ComponentModel;
+using System.Data;
 using System.Linq;
 using System.Windows;
 using SpawmetDatabase;
@@ -16,11 +17,13 @@ namespace SpawmetDatabaseWPF
         public event EventHandler<IEnumerable<Machine>> MachinesDeleted;
         public event EventHandler WorkCompleted;
 
+        public event EventHandler<Exception> ConnectionLost;
+
         //private readonly MachinesWindow _parentWindow;
         private readonly SpawmetDBContext _dbContext;
         private readonly IEnumerable<Machine> _machines;
 
-        private readonly BackgroundWorker _backgroundWorker;
+        private readonly BackgroundWorker _mainWorker;
         private readonly BackgroundWorker _initWorker;
 
         private int _deletedCount = 0;
@@ -36,43 +39,91 @@ namespace SpawmetDatabaseWPF
 
             WindowStartupLocation = WindowStartupLocation.CenterOwner;
 
+            this.ConnectionLost += ConnectionLostHandler;
+
             _initWorker = new BackgroundWorker();
-            _initWorker.DoWork += (sender, e) =>
+            _initWorker.WorkerSupportsCancellation = true;
+            _initWorker.DoWork += InitWorker_DoWork;
+            _initWorker.RunWorkerCompleted += InitWorker_RunWorkerCompleted;
+
+            _mainWorker = new BackgroundWorker();
+            _mainWorker.WorkerReportsProgress = true;
+            _mainWorker.WorkerSupportsCancellation = true;
+            _mainWorker.DoWork += MainWorker_DoWork;
+            _mainWorker.ProgressChanged += MainWorker_ProgressChanged;
+            _mainWorker.RunWorkerCompleted += MainWorker_RunWorkerCompleted;
+
+            this.Closed += (sender, e) =>
+            {
+                _initWorker.Dispose();
+                _mainWorker.Dispose();
+            };
+
+            _initWorker.RunWorkerAsync();
+        }
+
+        private void InitWorker_DoWork(object sender, DoWorkEventArgs e)
+        {
+            try
             {
                 foreach (var machine in _machines)
                 {
+                    if (_initWorker.CancellationPending)
+                    {
+                        return;
+                    }
+
                     _totalCount += machine.StandardPartSet.Count + machine.Orders.Count;
                     foreach (var order in machine.Orders)
                     {
                         _totalCount += order.AdditionalPartSet.Count;
                     }
                 }
+
+                if (_initWorker.CancellationPending)
+                {
+                    return;
+                }
                 _totalCount += _machines.Count();
-            };
-            _initWorker.RunWorkerCompleted += (sender, e) =>
+            }
+            catch (Exception exc)
             {
-                DeleteProgressBar.Minimum = 0;
-                DeleteProgressBar.Maximum = _totalCount;
-                DeleteProgressBar.Value = 0;
-                CounterTextBlock.Text = "0 z " + _totalCount;
+                Application.Current.Dispatcher.Invoke(() =>
+                {
+                    OnConnectionLost(exc);
+                });
+            }
+        }
 
-                TitleTextBlock.Text = "Usuwanie...";
+        private void InitWorker_RunWorkerCompleted(object sender, RunWorkerCompletedEventArgs e)
+        {
+            DeleteProgressBar.Minimum = 0;
+            DeleteProgressBar.Maximum = _totalCount;
+            DeleteProgressBar.Value = 0;
+            CounterTextBlock.Text = "0 z " + _totalCount;
 
-                _backgroundWorker.RunWorkerAsync();
-            };
-            
-            _backgroundWorker = new BackgroundWorker();
-            _backgroundWorker.WorkerReportsProgress = true;
-            _backgroundWorker.DoWork += (sender, e) =>
+            TitleTextBlock.Text = "Usuwanie...";
+
+            _mainWorker.RunWorkerAsync();
+        }
+
+        private void MainWorker_DoWork(object sender, DoWorkEventArgs e)
+        {
+            try
             {
                 foreach (var machine in _machines)
                 {
+                    if (_mainWorker.CancellationPending)
+                    {
+                        return;
+                    }
+
                     int count = machine.StandardPartSet.Count;
                     _dbContext.StandardPartSets.RemoveRange(machine.StandardPartSet);
                     _dbContext.SaveChanges();
 
                     _deletedCount += count;
-                    _backgroundWorker.ReportProgress(0);
+                    _mainWorker.ReportProgress(0);
 
                     foreach (var order in machine.Orders)
                     {
@@ -81,7 +132,7 @@ namespace SpawmetDatabaseWPF
                         _dbContext.SaveChanges();
 
                         _deletedCount += count;
-                        _backgroundWorker.ReportProgress(0);
+                        _mainWorker.ReportProgress(0);
                     }
 
                     count = machine.Orders.Count;
@@ -89,40 +140,64 @@ namespace SpawmetDatabaseWPF
                     _dbContext.SaveChanges();
 
                     _deletedCount += count;
-                    _backgroundWorker.ReportProgress(0);
+                    _mainWorker.ReportProgress(0);
                 }
-            };
-            _backgroundWorker.ProgressChanged += (sender, e) =>
+            }
+            catch (Exception exc)
             {
-                DeleteProgressBar.Value = _deletedCount;
-                CounterTextBlock.Text = _deletedCount + " z " + _totalCount;
-            };
-            _backgroundWorker.RunWorkerCompleted += (sender, e) =>
-            {
-                _deletedCount += _machines.Count();
+                Application.Current.Dispatcher.Invoke(() =>
+                {
+                    OnConnectionLost(exc);
+                });
+            }
+        }
 
-                _dbContext.Machines.RemoveRange(_machines);
+        private void MainWorker_ProgressChanged(object sender, ProgressChangedEventArgs e)
+        {
+            DeleteProgressBar.Value = _deletedCount;
+            CounterTextBlock.Text = _deletedCount + " z " + _totalCount;
+        }
+
+        private void MainWorker_RunWorkerCompleted(object sender, RunWorkerCompletedEventArgs e)
+        {
+            _deletedCount += _machines.Count();
+
+            _dbContext.Machines.RemoveRange(_machines);
+            try
+            {
                 _dbContext.SaveChanges();
-
-                OnMachinesDeleted(_machines);
-
-                DeleteProgressBar.Value += _deletedCount;
-                CounterTextBlock.Text = _deletedCount + " z " + _totalCount;
-
-                //parentWindow.Refresh();
-
-                OnWorkCompleted();
-
-                this.Close();
-            };
-
-            this.Closed += (sender, e) =>
+            }
+            catch (Exception exc)
             {
-                _initWorker.Dispose();
-                _backgroundWorker.Dispose();
-            };
+                OnConnectionLost(exc);
+                return;
+            }
 
-            _initWorker.RunWorkerAsync();
+            OnMachinesDeleted(_machines);
+
+            DeleteProgressBar.Value += _deletedCount;
+            CounterTextBlock.Text = _deletedCount + " z " + _totalCount;
+
+            //parentWindow.Refresh();
+
+            OnWorkCompleted();
+
+            this.Close();
+        }
+
+        private void ConnectionLostHandler(object sender, Exception exc)
+        {
+            _initWorker.CancelAsync();
+            _mainWorker.CancelAsync();
+
+            _initWorker.RunWorkerCompleted -= InitWorker_RunWorkerCompleted;
+            _mainWorker.DoWork -= MainWorker_DoWork;
+            _mainWorker.ProgressChanged -= MainWorker_ProgressChanged;
+            _mainWorker.RunWorkerCompleted -= MainWorker_RunWorkerCompleted;
+
+            this.Close();
+
+            this.ConnectionLost -= ConnectionLostHandler;
         }
 
         private void OnMachinesDeleted(IEnumerable<Machine> machines)
@@ -138,6 +213,14 @@ namespace SpawmetDatabaseWPF
             if (WorkCompleted != null)
             {
                 WorkCompleted(this, EventArgs.Empty);
+            }
+        }
+
+        private void OnConnectionLost(Exception exc)
+        {
+            if (ConnectionLost != null)
+            {
+                ConnectionLost(this, exc);
             }
         }
     }
